Restaurants : finalisation de la DAO

+++ model/DAO/RestaurantDAO.php
@@ -1,4 +1,5 @@
 <?php
+require_once 'model/domain/ExSimpleXMLElement.php';
 class RestaurantDAO
 {
   public $dataPath = 'data\restaurants.xml';
@@ -50,12 +51,17 @@ class RestaurantDAO
   public function addRestaurant($restaurantFilePath)
   {
     $restaurant = simplexml_load_file($restaurantFilePath);
-    $restaurant->addAttribute('id', uniqid());
-    $this->Restaurants->addChild($restaurant->getName(), $restaurant->asXML());
-    $this->Restaurants->asXML('data\\restau_temp.xml');
+    // generate NCNAME type id for the restaurant
+    $restaurant->addAttribute('id', "R" . strtoupper(uniqid()));
+    $restaurant->addAttribute('deleted', 'false');
+    $restaurants = new ExSimpleXMLElement($this->dataPath, 0, true);
+
+    // add all childs of the restaurant to the Restaurants xml
+    $restaurants->appendXML($restaurant);
+    $restaurants->asXML('data\\restau_temp.xml');
     $isValid = $this->isValid('data\\restau_temp.xml');
     if ($isValid[0]) {
-      $this->Restaurants->asXML($this->dataPath);
+      $restaurants->asXML($this->dataPath);
       unlink("data\\restau_temp.xml");
       return [true, null];
     } else {
@@ -63,21 +69,38 @@ class RestaurantDAO
     }
   }
 
-  public function updateRestaurant($restaurantFilePath)
+  public function updateRestaurant($restaurantFilePath, $id)
   {
+    // update the restaurant using the appendXML method
     $restaurant = simplexml_load_file($restaurantFilePath);
-    $restaurantId = $restaurant['id'];
-    $restaurantToUpdate = $this->Restaurants->xpath("//restaurant[@id='$restaurantId']")[0];
-    $restaurantToUpdate[0] = $restaurant->asXML();
-    $this->Restaurants->asXML('data\\restau_temp.xml');
-    $isValid = $this->isValid('data\\restau_temp.xml');
-    if ($isValid[0]) {
-      $this->Restaurants->asXML($this->dataPath);
-      unlink("data\\restau_temp.xml");
-      return [true, null];
+    $restaurant->addAttribute('id', $id);
+    $restaurant->addAttribute('deleted', 'false');
+    $restaurants = new ExSimpleXMLElement($this->dataPath, 0, true);
+    $restaurantToUpdate = $restaurants->xpath("//restaurant[@id='$id']")[0];
+    if ($restaurantToUpdate) {
+      // Mettre à jour les informations du restaurant
+      foreach ($restaurant->children() as $child) {
+        // Supprimer les anciens éléments avant d'ajouter les nouveaux
+        unset($restaurantToUpdate->{$child->getName()});
+        $restaurantToUpdate->appendXML($child);
+      }
+
+      // Sauvegarder les modifications dans un fichier temporaire
+      $restaurants->asXML('data\\restau_temp.xml');
+
+      $isValid = $this->isValid('data\\restau_temp.xml');
+      if ($isValid[0]) {
+        $restaurants->asXML($this->dataPath);
+        unlink('data\\restau_temp.xml');
+        return [true, null];
+      } else {
+        return [false, $isValid[1]];
+      }
     } else {
-      return [false, $isValid[1]];
+      return [false, "Restaurant non trouvé."];
     }
+
+
   }
 
   public function deleteRestaurant($id)
@@ -103,7 +126,8 @@ class RestaurantDAO
     $newRestaurant->adresse = $restaurant->coordonnees['adresse'];
     $newRestaurant->nomRestaurateur = (string) $restaurant->nom_restaurateur;
     $newRestaurant->setDescription($restaurant->description);
-
+    $newRestaurant->setCarte($restaurant->carte);
+    $newRestaurant->setMenus($restaurant->liste_menu, $restaurant->carte);
     return $newRestaurant;
   }
